Optimized packets
Packets were optimized using the previously implemented size computing 

Packets are now spread into two implementations LazyBufferPacket and ComposedPacket

LazyBufferPackets are those which are read from the network (a buffer) these packets have their content stored in the encoded format inside the stored buffer and the actual contents aren't computed unless they are specifically requested
when writing this type directly copies from its buffer

This is like the original implementation except theres now an additional type called ComposedPacket this implementation doesn't use a buffer and instead uses a list of Tdf's this implementation is used for when packets are created on the server. This removes the overhead of unknown sized buffer creation whenever a tdf builder is used. This implementation stores the tdfs in the list until writing time where in which the computeSize functions are called to create a buffer which perfectly fits the entire packet resulting in a packet that can be perfectly written to the output without having to reallocate / extend the size of the buffer it needs to be written into